\name{checkerspot}

\docType{data}

\alias{checkerspot}

\title{Bay Checkerspot Butterfly Population Data}

\description{
Estimated number of female Bay checkerspot butterflies (Euphydryas 
editha bayensis) from the JRC population 
(near San Francisco), each year from 1960 to 1986.  Abundances estimated with
mark-recapture methods.
}

\usage{
data(checkerspot)
}

\format{
A data frame with three columns:
\describe{
\item{yr}{Census year}
\item{N}{Estimated number of butterflies}
}
}

\source{
Morris and Doak (2002), table 4.1.  Data originally from Harrison et al. (1991).
}

\references{
Harrison, S., Quinn, J. F., Baughman, J. F., Murphy, D. D. and Erlich, P. R.
(1991) Estimating the effects of scientific study on two butterfly populations.
\emph{American Naturalist} \bold{137}, 227--243.

    Morris, W. F. and Doak D. F. (2002) \emph{Quantitative Conservation Biology:
        Theory and Practice of Population Viability Analysis}.
        Sunderland: Sinauer Associates.
}

\author{Bruce E. Kendall (\email{kendall@bren.ucsb.edu})}


\keyword{datasets}
