% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_sim_SEG.R
\name{demo_sim_SEG}
\alias{demo_sim_SEG}
\title{Stochastic Exponential Growth model demo}
\usage{
demo_sim_SEG()
}
\value{
Nothing is returned; the commmand is run for its side effect of 
launching a shiny app. You will have to close the shiny window or hit the
Stop button in the console to get the console prompt back.
}
\description{
A Shiny demo of the SEG model. The user can set the mean log growth rate,
the variance of the log growth rate, the initial population size, and the length of simulation.
The number of replicate simulations can also be changed.
}
\details{
Four graphs are produced. The left plots are time series of abundance, plotted on a linear 
scale (top) and a logarithmic scale (bottom). The right plots are histograms of the final 
abundance and final log abundance across the replicate simulations.

The simulations update immediately when you change any parameter value. To
see a new set of stochastic simulations without changing any parameters, hit 
the "Run again" button.
}
