% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_count_params.R
\name{estimate_SEG_params}
\alias{estimate_SEG_params}
\title{Estimate SEG model parameters}
\usage{
estimate_SEG_params(Nt, Yeart = seq(along = Nt), alpha = 0.05)
}
\arguments{
\item{Nt}{Time series of abundance. Must be a numeric vector of positive
values, with length at least 3}

\item{Yeart}{Sequence of times ("years") at which abundance is estimated
(must be the same length as \code{Nt}). If the abundance estimates are available
at each time step, this can be omitted}

\item{alpha}{Significance level for confidence levels (default is 0.05).}
}
\value{
A list with the following elements:

\item{mu}{Estimate of the mean growth rate}

\item{sigma2}{Estimate of the variance in the growth rate}

\item{params.CI}{A matrix with the CI of mu in the first row and the CI of
sigma2 in the second row}

\item{alpha}{The alpha level passed in by the user}
}
\description{
Given a time series of abundances, this function gives estimates and
confidence intervals for the parameters of the stochastic exponential growth
(SEG) parameters. It uses the "linear regression approach" described on pp.
120-121 of Dennis et al. (1991) (see also Chapter 3 of Morris & Doak 2002 for
a friendlier description). By default, the time series of abundances is
assumed to be equally spaced in time, with a timestep of one time unit;
deviations from this can be accomodated by providing a vector of times
associated with each abundance estimate.
}
\examples{
data(grizzly)
estimate_SEG_params(grizzly$N)
  
}
\references{
Dennis, B., Munholland, P. L. and Scott, J. M. (1991) Estimation
of growth and extinction parameters for endangered species.
\emph{Ecological  Monographs} \bold{61}, 115--143.

Morris, W. F. and Doak D. F. (2002) \emph{Quantitative Conservation
Biology: Theory and Practice of Population Viability Analysis}. Sunderland:
Sinauer Associates.
}
