% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRiskRicker.R
\name{extRiskRicker}
\alias{extRiskRicker}
\title{Calculate a CDF of quasi-extinction risk for the Ricker model}
\usage{
extRiskRicker(r, K, sigma2, time_horizon, N0, nsims = 1000, Nx = NULL,
  d = log(N0/Nx))
}
\arguments{
\item{r}{the low-density growth rate}

\item{K}{the equilibrium density}

\item{sigma2}{the variance of the log growth rates}

\item{time_horizon}{number of years to simulate}

\item{N0}{initial population size}

\item{nsims}{number of simulations}

\item{Nx}{quasi-extinction threshold}

\item{d}{the difference between \code{log(N0)} and \code{log(Nx)}. Either specify d or specify Nx}
}
\value{
A data frame with columns \code{Time} and \code{CDF}
Some error checking:
}
\description{
Calculate a CDF of quasi-extinction risk for the Ricker model
}

