\name{plot.ext.risk}

\alias{plot.ext.risk}

\title{Plot Method for ext.risk Objects}

\description{
}

\usage{
}

\arguments{
}

\details{
}

\value{
}

\references{
    Morris, W. F. and Doak D. F. (2002) \emph{Quantitative Conservation Biology:
        Theory and Practice of Population Viability Analysis}.
        Sunderland: Sinauer Associates.
}

\note{
}

\author{Bruce E. Kendall (\email{kendall@bren.ucsb.edu})}

\seealso{
}

\examples{
}

\keyword{hplot}
