\name{PVA-package}
\alias{PVA-package}
\alias{PVA}
\docType{package}
\title{
Population Viability Analysis
}
\description{
Perform population viability analysis (PVA) in R without needing advanced coding 
  techniques. The goal is for the tools to be accessible to any user who is comfortable with 
  standard data analysis in R. A GUI based on shiny will be included for many standard analyses.
}
\details{
\tabular{ll}{
Package: \tab PVA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-04\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Bruce Kendall
}
\references{

}
\keyword{ package }
\seealso{

}
\examples{
}
