\name{sealion}

\docType{data}

\alias{sealion}

\title{Northern Sea Lion Demographic Data}

\description{
Age-specific survival and fecundity of northern sea lion (\emph{Eumetopias
jubatus}) near Marmot Island, Alaska.
}

\usage{
data(sealion)
}

\format{
A data frame with three columns:
\describe{
\item{age}{Age in years}
\item{surv}{Annual survival probability}
\item{fec}{Annual average fecundity}
}
The data in the last row applies to ages 13 and above; the oldest individual 
observed in this population was 31 years old.
}

\source{
Ebert (1999), table 6.8.  Data originally from Calkins and Pitcher (1982).
}

\references{
Calkins, D. G. and Pitcher, K. W. (1982)  \emph{Population Assessment,
Ecology, and Trophic Relationships of Stellar Sea Lions in the Gulf of Alaska}.
Alaska Department of Fish and Game Final Report RU243.  Anchorage: Alaska
Department of Fish and Game.

Ebert, T. A. (1999) \emph{Plant and Animal Populations: Methods in Demography}.
San Diego: Academic Press.
}

\author{Bruce E. Kendall (\email{kendall@bren.ucsb.edu})}

\keyword{datasets}
