% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRicker.R
\name{simulateRicker}
\alias{simulateRicker}
\title{Simulate one or more time series from the stochastic Ricker model}
\usage{
simulateRicker(r, K, sigma2, N0, time_horizon, nsims = 1, keep_N0 = FALSE)
}
\arguments{
\item{r}{the low-density growth rate}

\item{K}{the equilibrium density}

\item{sigma2}{the variance of the log growth rates}

\item{N0}{initial population size}

\item{time_horizon}{number of years to simulate}

\item{nsims}{number of simulations}

\item{keep_N0}{(logical) whether to return \code{N0} at the beginning of the time series}
}
\value{
If \code{nsim=1} (the default), a vector of length \code{time_horizon} (or \code{time_horizon+1} if \code{keep_N0=TRUE}) with the simulated abundances. Otherwise, a matrix with \code{time_horizon} (or \code{time_horizon+1} if \code{keep_N0=TRUE}) rows and \code{nsim} columns.
}
\description{
Simulate one or more time series from the stochastic Ricker model
}
\details{
This function simulates the Ricker discrete time model with environmental stochasticity, defined as follows
\deqn{N(t+1) = N(t) exp[r (1 - N(t)/K) + \epsilon(t)]},
where
\deqn{\epsilon(t) ~ Normal(0, \sigma)}
}
\examples{
simulateRicker(2, 100, 0.01, 20, 10)
simulateRicker(2, 100, 0.01, 20, 10, 5, keep_N0=TRUE)
}
