% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetDB2.R
\name{subsetDB2}
\alias{subsetDB2}
\title{Extract subset from a compadre/comadre object}
\usage{
subsetDB2(sub, db = comadre)
}
\arguments{
\item{sub}{Logical expression that can be used ot subset the metadata component of db}

\item{db}{A Compadre or Comadre object}
}
\value{
A com(p)adre object (list), with each row of the returned metadata being matched by an element in the other components
}
\description{
Extract subset from a compadre/comadre object
}
\details{
Based on subsetDB() from https://github.com/jonesor/compadreDB; the change is to allow the subsetting to be specified as a logical expression
}
