% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrstats.R
\name{vrstats}
\alias{vrstats}
\title{Calculate vital rate stats for use in vitalsim}
\usage{
vrstats(vrmat)
}
\arguments{
\item{vrmat}{A matrix of vital rates, with one year per column; typically the output from \code{\link{extractVRs}}}
}
\value{
A list with elements vrmeans, vrvars, corrin, and corrout, which are used by vitalsim() in the popbio package
}
\description{
Calculate vital rate stats for use in vitalsim
}

