% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrtomat.R
\name{vrtomat}
\alias{vrtomat}
\title{Reconstitute matrix from vital rates}
\usage{
vrtomat(vrs)
}
\arguments{
\item{vrs}{A vector or 1-column matrix of vital rates}
}
\value{
A projection matrix
}
\description{
Reconstitute matrix from vital rates
}
\details{
An object called \code{matstruct} (part of the output from \code{\link{extractVRs}}) needs to be in the global environment. This is not passed in because \code{vitalsim} does not allow for the passing of extra parameters through to the matrix construction function.
}
